# api_spec.yaml
openapi: 3.0.3
info:
  title: 课堂行为分析系统 API
  description: |
    基于VHEAT算法的学生课堂行为智能识别系统API文档
    
    ## 功能模块
    1. 账号管理模块 - 用户注册、登录、权限管理
    2. 视频处理模块 - 视频上传、处理状态查询
    3. 行为分析模块 - 行为统计、时间序列、异常检测
    4. 教学优化模块 - 优化建议、跨课次对比
    5. 报告导出模块 - PDF报告生成
    
    ## 认证方式
    使用JWT Bearer Token认证，在请求头中添加：
    ```
    Authorization: Bearer {token}
    ```
  version: 1.0.0
  contact:
    name: 系统开发团队
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发环境
  - url: https://api.example.com/api/v1
    description: 生产环境
...
paths:
...
  /optimization/{video_id}/recommendations:
    get:
      tags:
        - 教学优化
      summary: 获取优化建议
      description: |
        基于行为分析结果生成教学优化建议。
        包括互动改进、注意力提升、节奏调整等建议。
      security:
        - bearerAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_id:
                    type: integer
                    example: 1
                  recommendations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
  /optimization/compare:
    get:
      tags:
        - 教学优化
      summary: 跨课次对比
      description: |
        对比多个课次的行为指标，分析教学效果变化趋势。
        支持变点检测和趋势分析。
      security:
        - bearerAuth: []
      parameters:
        - name: video_ids
          in: query
          required: true
          schema:
            type: string
          description: 视频ID列表，逗号分隔，如 "1,2,3"
        - name: metrics
          in: query
          schema:
            type: string
            default: "all"
          description: 对比指标，可选值：all, interaction, attention, engagement
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonResult'
  /optimization/{video_id}/highlights:
    get:
      tags:
        - 教学优化
      summary: 获取优秀课堂片段
      description: |
        根据预设指标阈值自动筛选高质量课堂片段。
        包括互动占比高、异常行为少、师生互动良好的时段。
      security:
        - bearerAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: integer
        - name: min_score
          in: query
          schema:
            type: number
            format: float
            default: 0.8
          description: 最低质量评分（0-1）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_id:
                    type: integer
                    example: 1
                  highlights:
                    type: array
                    items:
                      $ref: '#/components/schemas/Highlight'

  # ==================== 报告导出模块 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token认证
  schemas:
...
    Recommendation:
      type: object
      properties:
        type:
          type: string
          enum: [interaction, attention, rhythm, engagement]
          example: "interaction"
          description: 建议类型
        priority:
          type: string
          enum: [low, medium, high]
          example: "high"
          description: 优先级
        title:
          type: string
          example: "增加互动频次"
        description:
          type: string
          example: "讨论占比不足（当前15%，建议≥30%），建议增加分组讨论环节"
        suggested_actions:
          type: array
          items:
            type: string
          example:
            - "在课程中段（20-30分钟）增加小组讨论"
            - "设置讨论问题，引导学生互动"
    ComparisonResult:
      type: object
      properties:
        comparison:
          type: array
          items:
            type: object
            properties:
              video_id:
                type: integer
                example: 1
              lesson_date:
                type: string
                format: date
                example: "2025-10-20"
              metrics:
                type: object
                properties:
                  interaction_rate:
                    type: number
                    format: float
                    example: 0.25
                    description: 互动率（0-1）
                  attention_rate:
                    type: number
                    format: float
                    example: 0.85
                    description: 注意力率（0-1）
                  engagement_score:
                    type: number
                    format: float
                    example: 0.75
                    description: 参与度评分（0-1）
        trends:
          type: object
          properties:
            interaction_rate:
              type: string
              enum: [increasing, decreasing, stable]
              example: "increasing"
            attention_rate:
              type: string
              enum: [increasing, decreasing, stable]
              example: "stable"
    Highlight:
      type: object
      properties:
        start_time:
          type: integer
          example: 600
          description: 片段开始时间（秒）
        end_time:
          type: integer
          example: 900
          description: 片段结束时间（秒）
        score:
          type: number
          format: float
          example: 0.92
          description: 质量评分（0-1）
        reasons:
          type: array
          items:
            type: string
          example:
            - "互动占比高（35%）"
            - "异常行为少（低头率<10%）"
            - "教师引导与学生响应良好"
        thumbnail_url:
          type: string
          example: "/api/v1/videos/1/frames/600"
          description: 缩略图URL
...